refactor(TableGroup): OrderTable 연관관계 매핑
