+++ src/main/java/kitchenpos/tablegroup/TableGroupService.java
@@ -8,7 +8,6 @@ import kitchenpos.ui.dto.TableGroupRequest;
 import kitchenpos.ui.dto.TableGroupResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.CollectionUtils;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -28,31 +27,17 @@ public class TableGroupService {
 
     @Transactional
     public TableGroupResponse create(final TableGroupRequest tableGroup) {
-        List<OrderTable> orderTables = toOrderTables(tableGroup.getOrderTables());
-        validateSizeOf(orderTables);
         TableGroup newTableGroup = tableGroupRepository.save(new TableGroup());
-        orderTables.forEach(table -> table.assign(newTableGroup.getId()));
 
-        return TableGroupResponse.from(newTableGroup, orderTables);
-    }
+        newTableGroup.add(toOrderTables(tableGroup.getOrderTables()));
 
-    private void validateSizeOf(List<OrderTable> orderTables) {
-        /*private static */
-        final int MINIMUM_GROUP_SIZE = 2;
-        if (CollectionUtils.isEmpty(orderTables) || orderTables.size() < MINIMUM_GROUP_SIZE) {
-            throw new IllegalArgumentException(MINIMUM_GROUP_SIZE + "개 이상의 테이블이 필요합니다");
-        }
+        return TableGroupResponse.from(newTableGroup, toOrderTables(tableGroup.getOrderTables()));
     }
 
     @Transactional
     public void ungroup(final Long tableGroupId) {
-        tableGroupRepository.validateContains(tableGroupId);
-
-        List<OrderTable> orderTables = orderTableRepository.findAllByTableGroupId(tableGroupId);
-        for (OrderTable orderTable : orderTables) {
-            noOngoingOrderValidator.validate(orderTable);
-            orderTable.ungroup();
-        }
+        TableGroup tableGroup = tableGroupRepository.getBy(tableGroupId);
+        tableGroup.ungroupUsing(noOngoingOrderValidator);
     }
 
     private List<OrderTable> toOrderTables(List<OrderTableRequest> orderTableRequests) {
